apps/api/requirements.txt
# FastAPI
fastapi==0.121.3
uvicorn[standard]==0.38.0
python-dotenv==1.2.1

# Database - PostgreSQL (RDS)
SQLAlchemy==2.0.44
asyncpg==0.30.0
psycopg2-binary==2.9.11
alembic==1.17.2

# AWS SDK
boto3>=1.34.0
aioboto3==13.1.1
aws-xray-sdk>=2.14.0

# Cache - Redis
redis==7.1.0

# Search - OpenSearch
opensearch-py==2.7.1

# Auth - JWT
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
PyJWT==2.10.0

# Validation
pydantic==2.12.4
pydantic-settings>=2.2.0
email-validator==2.3.0

# Date utilities
python-dateutil>=2.8.0

# GraphQL
strawberry-graphql[fastapi]==0.291.3

# HTTP Client
httpx==0.28.1

# Logging
loguru==0.7.3

# Rate Limit
slowapi==0.1.9

# Dev tools
black
ruff
mypy
pytest
